BhaagLagam/requirements.txt
django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pynamodb==6.0.2
boto3==1.34.51
python-dotenv==1.0.1
whitenoise==6.6.0
Pillow==10.2.0
django-storages==1.14.2